command-center/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

APIFlask==2.1.1
Authlib==1.3.1
Faker==26.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.29
